Added extra fields requested by the service [not working]
making a commit only to share the changes. Thing is this should create a bunch of typescript errors due to us referencing the non-existing fields as far as the 'Asset' model is concerned. For these changes to work, you'll either need to update the node_module locally, or release a new version of the mtfh-frontend-common that has all the updates to the models.